+++ odw/core/etl/transformation/standardised/appeal_attribute_matrix_standardisation_process.py
@@ -3,6 +3,7 @@ from typing import Any
 
 
 class AppealAttributeMatrixStandardisationProcess(StandardisationProcess):
+    CSV_FOLDER = "AppealAttributeMatrix"
     OUTPUT_TABLE = "appeal_attribute_matrix"
 
     def __init__(self, spark):
